Revert "Made suggested changes"
