+++ src/App.js
@@ -1,15 +1,15 @@
 import "./App.css";
 import { createContext, useState } from "react";
-import Login from "./authentication/Login";
-import NavBar from "./navigation/Navbar";
-import AllProducts from "./products/retrieveProducts/AllProducts";
-import { ACCESS_TOKEN } from "./constants/Constants";
+import Login from "./Components/login";
+import allProducts from "./Components/allProducts";
+import NavBar from "./Components/Navbar";
+import AllProducts from "./Components/allProducts";
 
 export const AppContext = createContext();
 
 function App() {
   const [loggedIn, setLoggedIn] = useState(() => {
-    const token = localStorage.getItem(ACCESS_TOKEN);
+    const token = localStorage.getItem("accessToken");
     return token != null;
   });
 
@@ -17,10 +17,10 @@ function App() {
 
   return (
     <AppContext.Provider value={{ setLoggedIn, openModal, setOpenModal }}>
-      {!loggedIn && <Login setLoggedIn={setLoggedIn} />}
+      {!loggedIn && <Login />}
       {loggedIn && (
         <>
-          <NavBar setLoggedIn={setLoggedIn} />
+          <NavBar />
           <AllProducts />
         </>
       )}
