+++ src/Components/Navbar.jsx
@@ -7,25 +7,26 @@ import Avatar from "@mui/material/Avatar";
 import MenuItem from "@mui/material/MenuItem";
 import Menu from "@mui/material/Menu";
 import CircularProgress from "@mui/material/CircularProgress";
-import { useState } from "react";
+import Button from "@mui/material/Button";
+import ProductDetails from "./productDetails";
+import { useState, useContext } from "react";
+import { AppContext } from "../App";
 import { useQuery } from "react-query";
-import { getUserProfile } from "../axios/ApiCalls";
-import AddProduct from "../products/createProducts/AddProduct";
-import { ACCESS_TOKEN, KEYS, REFRESH_TOKEN } from "../constants/Constants";
+import { getUserProfile } from "../apiCalls";
 
-export default function Navbar({ setLoggedIn }) {
+export default function Navbar() {
   const [anchorEl, setAnchorEl] = useState(null);
-  const [openModal, setOpenModal] = useState(false);
+  const { setLoggedIn, openModal, setOpenModal } = useContext(AppContext);
 
   const query = useQuery({
-    queryKey: KEYS.GET_PROFILE,
-    queryFn: () => getUserProfile(localStorage.getItem(ACCESS_TOKEN)),
+    queryKey: "getProfile",
+    queryFn: () => getUserProfile(localStorage.getItem("accessToken")),
   });
 
   const handleLogout = () => {
     handleClose();
-    localStorage.removeItem(ACCESS_TOKEN);
-    localStorage.removeItem(REFRESH_TOKEN);
+    localStorage.removeItem("accessToken");
+    localStorage.removeItem("refreshToken");
     setLoggedIn(false);
   };
 
@@ -35,6 +36,10 @@ export default function Navbar({ setLoggedIn }) {
 
   const handleClose = () => {
     setAnchorEl(null);
+  };
+
+  const handleAddProduct = () => {
+    setOpenModal(true);
   };
 
   return (
@@ -49,7 +54,13 @@ export default function Navbar({ setLoggedIn }) {
             {query.isLoading && <CircularProgress color="secondary" />}
             {!query.isLoading && (
               <div>
-                <AddProduct openModal={openModal} setOpenModal={setOpenModal} />
+                <Button
+                  variant="outlined"
+                  color="inherit"
+                  onClick={handleAddProduct}
+                >
+                  Add Product
+                </Button>
                 <IconButton
                   size="large"
                   aria-label="account of current user"
@@ -82,6 +93,20 @@ export default function Navbar({ setLoggedIn }) {
           </Toolbar>
         </AppBar>
       </Box>
+      {openModal && (
+        <ProductDetails
+          openModal={openModal}
+          toAdd={true}
+          setOpenModal={setOpenModal}
+          product={{
+            title: "",
+            description: "",
+            price: "",
+            category: { id: "" },
+            image: "",
+          }}
+        />
+      )}
     </>
   );
 }
